transport/net.cc: widen buffer size/offset to ssize_t to fix integer overflow

Mirrored-from: 645c44e05374e134c0aaba98366700b0672546ec

+++ src/include/collectives.h
@@ -71,7 +71,7 @@ inline int ncclTypeSize(ncclDataType_t type) {
 #define NCCL_MODE_PTR 2
 struct ncclConnFifo {
   int mode;
-  int offset;
+  ssize_t offset;
   ssize_t size;
   void* ptr;
 };
